@@ -9,24 +9,28 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  audioplayers: ^0.16.1
-  cupertino_icons: ^0.1.3
-  easy_localization: ^2.3.3
-  equatable: ^1.2.5
-  firebase_auth: ^0.18.0+1
-  firebase_core: ^0.5.0
-  flutter_dotenv: ^2.1.0
-  flutter_login_facebook: ^0.2.1
+  audioplayers: ^0.17.4
+  cupertino_icons: ^1.0.2
+  easy_localization: ^3.0.0
+  equatable: ^2.0.0
+  firebase_auth: ^1.0.0
+  firebase_core: ^1.0.0
+  flutter_dotenv: ^4.0.0-nullsafety.0
+  flutter_login_facebook: ^1.0.0-nullsafety.1
   flutter_spinkit: ^4.1.2+1
-  flutter_svg: any
-  fluttertoast: ^7.1.1
-  font_awesome_flutter: ^8.8.1
-  google_sign_in: ^4.5.3
+  flutter_svg: ^0.19.3
+  fluttertoast: ^7.1.8
+  font_awesome_flutter: ^8.11.0
+  google_sign_in: ^5.0.0
   hive: ^1.4.4+1
-  intl: ^0.16.1
+  http: ^0.13.0
+  intl: ^0.17.0
   modal_progress_hud: ^0.1.3
-  path_provider: ^1.6.18
-  provider: ^4.3.2+2
+  path_provider: ^2.0.1
+  provider: ^4.3.3
+
+dependency_overrides:
+  path_provider: any
 
 dev_dependencies:
   flutter_test:
